refactor: factor out `searchFiltersFactory`
